+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 6.0.1 (pending)
 
+* log validation report
 * dumped memory quota handling, kubernetes is responsible for that
 * dumped environment field - it's too container centric now that we have kubernetes; use 'sc ssh' instead
 * server startup output with image version